Make sure the output str is one time use
Previously, the output string was an attribute of the object which was
not ideal since the string will keep changing for different messages as
well as the string is not used by anywhere outside the library.

Does, changed it from that to being a string that is returned directly
from the method.

Also, it was causing a bug. Since the string was being appended to, the
messages were getting added after the string.