Fix issue with finding caller details
The details of the caller are found by using sys. However, while using
it we need to trace back to the caller which is one layer before the
caller method.

Previously the caller was being extracted in the Formatter class,
however that was causing issues, such as it added a lot many layers to
the stack so it became hard to find the actual caller. Thus the caller
is now extracted in the logger class itself which makes it easier with
the disadvantage of passing that frame to a few methods before
formatting it to the string, however that overhead is worth it.