Allow dir and file for log_path
The log_path param will now accept a directory as well as a file both.
If it would be a directory, a log file will be created inside that
directory and accordingly logs will be written to it.

If it is a file, logs will be written to that file.