Make some logical changes to create file stream
The file stream should not be created only if the log file path is
invalid. Even if the `disable_file` flag is passed it should be
created because the `disable_file` can be altered later by the user
in which case we need to resume writing to the file.

However, if the log_file was invalid during init and still the
`disable_file` flag is passed, it won't have any effect.