Fix bug of get_log_file returning None
Since the instances are handled by the latest updated instance, we
cannot just return the name of the log file when the `get_log_file`
method is called. This is because all the instances share the streams
and so a new stream might be added by some other instances and the rest
of the instances will not have the `log_file` name updated.

Thus, we need to go through each stream and find out which ones are
file streams and accordingly return a list of those.